+++ gql_dedupe_link/example/gql_dedupe_link_example.dart
@@ -1,5 +1,5 @@
-import "package:gql/link.dart";
 import "package:gql_dedupe_link/gql_dedupe_link.dart";
+import "package:gql_link/gql_link.dart";
 
 void main() {
   // ignore: unused_local_variable