+++ gql_link/lib/gql_link.dart
@@ -1,0 +1,7 @@
+/// [DocumentNode]-based GraphQL execution interface.
+library link;
+
+export "package:gql_link/src/exceptions.dart";
+export "package:gql_link/src/link.dart";
+export "package:gql_link/src/request_serializer.dart";
+export "package:gql_link/src/response_parser.dart";