+++ gql_link/lib/src/exceptions.dart
@@ -1,5 +1,4 @@
 import "package:gql/execution.dart";
-import "package:http/http.dart" as http;
 import "package:meta/meta.dart";
 
 /// Exception occurring when serialization fails.
@@ -19,12 +18,8 @@ class ParserException implements Exception {
   /// The original exception causing this exception
   final Exception originalException;
 
-  /// Response which caused the exception
-  final http.Response response;
-
   const ParserException({
     @required this.originalException,
-    @required this.response,
   });
 }
 
@@ -32,14 +27,10 @@ class ParserException implements Exception {
 /// or parsed response is missing both `data` and `errors`.
 @immutable
 class ServerException implements Exception {
-  /// Response which caused the exception
-  final http.Response response;
-
   /// The parsed response
   final Response parsedResponse;
 
   const ServerException({
-    @required this.response,
     @required this.parsedResponse,
   });
 }