+++ gql_code_gen/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.2
+
+- Upgrade `package:gql` to v0.11.0
+
 ## 0.1.1
 
 - Upgrade `package:gql` to v0.10.0